added state bucket, empty environment platform module

+++ bin/setup_development_environment.sh
@@ -63,27 +63,34 @@ install_hadolint() {
     arch_code="x86_64"
   fi
 
-  download_url="$(curl -s https://api.github.com/repos/hadolint/hadolint/releases/latest | grep Linux | grep browser_download_url | grep -v sha256 | grep "${arch_code}" | awk '{gsub("\"", ""); print $2}')"
+  if ! command -v hadolint >>/dev/null; then
+    download_url="$(curl -s https://api.github.com/repos/hadolint/hadolint/releases/latest | grep Linux | grep browser_download_url | grep -v sha256 | grep "${arch_code}" | awk '{gsub("\"", ""); print $2}')"
 
-  echo "Hadolint download url: ${download_url}" 1>&2
+    echo "Hadolint download url: ${download_url}" 1>&2
 
-  curl -L -o hadolint-test "${download_url}"
-  ${SUPERUSER_COMMAND} mv hadolint-test /usr/local/bin/hadolint
-  ${SUPERUSER_COMMAND} chmod +x /usr/local/bin/hadolint
+    curl -L -o hadolint-test "${download_url}"
+    ${SUPERUSER_COMMAND} mv hadolint-test /usr/local/bin/hadolint
+    ${SUPERUSER_COMMAND} chmod +x /usr/local/bin/hadolint
+  fi
+
 }
 
 install_tfenv() {
-  git clone --depth=1 https://github.com/tfutils/tfenv.git ~/.tfenv
-  ln -s ~/.tfenv/bin/* /usr/local/bin
+  if ! command -v tfenv >>/dev/null; then
+    git clone --depth=1 https://github.com/tfutils/tfenv.git ~/.tfenv
+    ln -s ~/.tfenv/bin/* /usr/local/bin
+  fi
 
   tfenv install latest
 }
 
 install_aws_cli() {
-  curl "https://awscli.amazonaws.com/awscli-exe-linux-aarch64.zip" -o "awscliv2.zip"
-  unzip awscliv2.zip
-  ${SUPERUSER_COMMAND} ./aws/install
-  rm -rf ./aws ./awscliv2.zip
+  if ! command -v aws >>/dev/null; then
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-aarch64.zip" -o "awscliv2.zip"
+    unzip awscliv2.zip
+    ${SUPERUSER_COMMAND} ./aws/install
+    rm -rf ./aws ./awscliv2.zip
+  fi
 }
 
 install_kubectl() {
@@ -93,17 +100,44 @@ install_kubectl() {
     arch_code="amd64"
   fi
 
-  curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/${arch_code}/kubectl"
+  if ! command -v kubectl >>/dev/null; then
+    curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/${arch_code}/kubectl"
 
-  ${SUPERUSER_COMMAND} install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
-  rm -f ./kubectl
+    ${SUPERUSER_COMMAND} install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
+    rm -f ./kubectl
+  fi
 }
 
 install_helm() {
-  curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
-  chmod 700 get_helm.sh
-  ${SUPERUSER_COMMAND} ./get_helm.sh
-  rm ./get_helm.sh
+  if ! command -v helm >>/dev/null; then
+    curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
+    chmod 700 get_helm.sh
+    export VERIFY_CHECKSUM=false
+    ${SUPERUSER_COMMAND} ./get_helm.sh
+    unset VERIFY_CHECKSUM
+    rm ./get_helm.sh
+  fi
+}
+
+install_pyenv() {
+  if ! command -v pyenv >>/dev/null; then
+    curl https://pyenv.run | bash
+  fi
+
+  export PYENV_ROOT="$HOME/.pyenv"
+  # shellcheck disable=SC3010
+  [ -d $PYENV_ROOT/bin ] && export PATH="$PYENV_ROOT/bin:$PATH"
+  eval "$(pyenv init -)"
+
+  pyenv install -s 3.13.0
+
+  if [ "$(pyenv global)" = "system" ]; then
+    pyenv global 3.13.0
+  fi
+}
+
+install_pre_commit() {
+  python3 -m pip install -U pre-commit
 }
 
 install_with_brew() {
@@ -121,6 +155,8 @@ install_with_apt_get() {
   install_aws_cli
   install_kubectl
   install_helm
+  install_pyenv
+  install_pre_commit
 }
 
 install_with_dnf() {
@@ -134,6 +170,8 @@ install_with_dnf() {
   install_aws_cli
   install_kubectl
   install_helm
+  install_pyenv
+  install_pre_commit
 }
 
 main() {
@@ -145,6 +183,8 @@ main() {
 
   # shellcheck disable=SC2091
   $(echo "install_with_${package_manager}" | awk '{gsub("-", "_"); print $0;}')
+
+  echo "Installations complete, make sure to add any necessary changes to shell configurations as needed."
 }
 
 # shellcheck disable=SC2068
